chore: create MRT local provider
MRT: Material React Table
Create a local provider to prevent re-renders in table, states.
