+++ src/index.jsx
@@ -13,6 +13,7 @@ import { AdapterDateFns } from '@mui/x-date-pickers/AdapterDateFns';
 import { es } from 'date-fns/locale';
 import { LoaderProvider } from '@providers/loader';
 import { UsersProvider } from '@providers/users';
+import { LeadsMRTProvider } from '@providers/local/LeadsMRT/provider';
 
 const container = document.getElementById('root');
 const root = createRoot(container);
@@ -26,9 +27,11 @@ root.render(
 					<LoadingProvider>
 						<SessionProvider>
 							<UsersProvider>
-								<Router basename='/'>
-									<App />
-								</Router>
+								<LeadsMRTProvider>
+									<Router basename='/'>
+										<App />
+									</Router>
+								</LeadsMRTProvider>
 							</UsersProvider>
 						</SessionProvider>
 					</LoadingProvider>
