+++ src/views/main/leads/useLeads.js
@@ -1,8 +1,9 @@
 import { endOfMonth, startOfMonth } from 'date-fns';
 import { useLoaderContext } from '@providers/loader';
-import { useQuery } from '@apollo/client';
+import { useMutation, useQuery } from '@apollo/client';
 import { GET_REQUESTS } from '@views/main/requests/queryRequests';
 import { useEffect, useState } from 'react';
+import { useLeadsMRTContext } from '@providers/local/LeadsMRT/provider';
 
 const useLeads = () => {
 	const loader = useLoaderContext();
@@ -10,63 +11,18 @@ const useLeads = () => {
 	const closeFilterMenu = () => setFilterMenuVisible(false);
 	const openFilterMenu = () => setFilterMenuVisible(true);
 
-	// Settings States
-	// Global filters for useQuery
-	const [selectedSeller, setSelectedSeller] = useState(
-		JSON.parse(sessionStorage.getItem('seller')) ?? {}
-	);
-
-	const [dateRange, setDateRange] = useState([
-		{
-			startDate: startOfMonth(new Date()),
-			endDate: endOfMonth(new Date()),
-			key: 'selection',
-		},
-	]);
-
-	// Muestra las propias o todas
-	const [showAll, setShowAll] = useState(
-		JSON.parse(sessionStorage.getItem('show')) ?? true
-	);
-
-	const [showPending, setShowPending] = useState(
-		JSON.parse(sessionStorage.getItem('pending')) ?? false
-	);
-
-	const [columnSizeModel, setColumnSizeModel] = useState({});
-
-	const [columnVisibilityModel, setColumnVisibilityModel] = useState({
-		extraComments: false,
-		updatedAt: false,
-		createdAt: false,
-	});
-
-	const [columnOrderModel, setColumnOrderModel] = useState([
-		'requestStatusIcon',
-		'requestStatus',
-		'shortId',
-		'requestDate',
-		'assignedUser',
-		'brand',
-		'serviceType',
-		'contactMedium',
-		'advertisingMedium',
-		'productStatus',
-		'comments',
-		'company',
-		'clientName',
-		'clientPhoneNumber',
-		'extraComments',
-		'createdAt',
-		'updatedAt',
-		'createdBy',
-		'isSale',
-	]);
-
-	const [paginationModel, setPaginationModel] = useState({
-		pageIndex: 0,
-		pageSize: 30,
-	});
+	const {
+		showAll,
+		setShowAll,
+		selectedSeller,
+		setSelectedSeller,
+		dateRange,
+		setDateRange,
+		showPending,
+		setShowPending,
+		paginationModel,
+		setCountRows,
+	} = useLeadsMRTContext();
 
 	const { data, error, loading, refetch } = useQuery(GET_REQUESTS, {
 		variables: {
@@ -85,108 +41,15 @@ const useLeads = () => {
 	});
 
 	useEffect(() => {
-		const items = JSON.parse(sessionStorage.getItem('date'));
+		if (data) {
+			const aux = data.requests.results;
+			const auxCount = data.requests.info.count;
 
-		if (items) {
-			setDateRange([
-				{
-					startDate: new Date(
-						JSON.parse(sessionStorage.getItem('date'))[0].startDate
-					),
-					endDate: new Date(
-						JSON.parse(sessionStorage.getItem('date'))[0].endDate
-					),
-					key: 'selection',
-				},
-			]);
+			console.log('Requests: ', aux);
+			console.log('auxCount', auxCount);
+			setCountRows(auxCount);
 		}
-	}, []);
-
-	useEffect(() => {
-		const items = JSON.parse(sessionStorage.getItem('seller'));
-		if (items) {
-			setSelectedSeller(items);
-		}
-	}, []);
-
-	useEffect(() => {
-		const items = JSON.parse(sessionStorage.getItem('show'));
-		if (items) {
-			setShowAll(items);
-		}
-	}, []);
-
-	useEffect(() => {
-		const items = JSON.parse(sessionStorage.getItem('pending'));
-		if (items) {
-			setShowPending(items);
-		}
-	}, []);
-
-	useEffect(() => {
-		const items = JSON.parse(localStorage.getItem('headersVisibility'));
-		if (items) {
-			setColumnVisibilityModel(items);
-		}
-	}, []);
-
-	useEffect(() => {
-		const items = JSON.parse(localStorage.getItem('headersOrder'));
-		if (items) {
-			setColumnOrderModel(items);
-		}
-	}, []);
-
-	useEffect(() => {
-		const items = JSON.parse(localStorage.getItem('headersSize'));
-		if (items) {
-			setColumnSizeModel(items);
-		}
-	}, []);
-
-	useEffect(() => {
-		const items = JSON.parse(localStorage.getItem('pageSize'));
-		if (items) {
-			setPaginationModel({ pageIndex: 0, pageSize: items?.pageSize });
-		}
-	}, []);
-
-	// Save props on change
-
-	useEffect(() => {
-		localStorage.setItem(
-			'headersVisibility',
-			JSON.stringify(columnVisibilityModel)
-		);
-	}, [columnVisibilityModel]);
-
-	useEffect(() => {
-		localStorage.setItem('headersOrder', JSON.stringify(columnOrderModel));
-	}, [columnOrderModel]);
-
-	useEffect(() => {
-		localStorage.setItem('headersSize', JSON.stringify(columnSizeModel));
-	}, [columnSizeModel]);
-
-	useEffect(() => {
-		localStorage.setItem('pageSize', JSON.stringify(paginationModel));
-	}, [paginationModel]);
-
-	useEffect(() => {
-		sessionStorage.setItem('date', JSON.stringify(dateRange));
-	}, [dateRange]);
-
-	useEffect(() => {
-		sessionStorage.setItem('pending', JSON.stringify(showPending));
-	}, [showPending]);
-
-	useEffect(() => {
-		sessionStorage.setItem('show', JSON.stringify(showAll));
-	}, [showAll]);
-
-	useEffect(() => {
-		sessionStorage.setItem('seller', JSON.stringify(selectedSeller));
-	}, [selectedSeller]);
+	}, [data]);
 
 	return {
 		allLeads: { showAll, setShowAll },
@@ -204,15 +67,6 @@ const useLeads = () => {
 			loading,
 		},
 		pending: { showPending, setShowPending },
-		table: {
-			columnOrder: { model: columnOrderModel, onChange: setColumnOrderModel },
-			columnSize: { model: columnSizeModel, onChange: setColumnSizeModel },
-			columnVisibility: {
-				model: columnVisibilityModel,
-				onChange: setColumnVisibilityModel,
-			},
-			pagination: { model: paginationModel, onChange: setPaginationModel },
-		},
 	};
 };
 export default useLeads;
